Reogranize deopt info panels in preparation for map explorer panel
