+++ packages/v8-deopt-webapp/src/components/V8DeoptInfoPanel/DeoptTables.scss
@@ -1,23 +1,3 @@
-.deoptsListPanel {
-	border-top-left-radius: 0;
-	overflow: auto;
-
-	.panel_title {
-		font-size: 1rem;
-		font-weight: 500;
-		text-align: center;
-		margin-bottom: 0;
-	}
-
-	.tabLink {
-		white-space: nowrap;
-	}
-
-	> nav {
-		box-shadow: 0 0 0 black;
-	}
-}
-
 .entryTable {
 	display: block;
 	margin: 1rem auto;
