+++ packages/v8-deopt-webapp/src/components/V8DeoptInfoPanel/index.scss
@@ -1,0 +1,19 @@
+.v8deoptInfoPanel {
+	border-top-left-radius: 0;
+	overflow: auto;
+
+	.panel_title {
+		font-size: 1rem;
+		font-weight: 500;
+		text-align: center;
+		margin-bottom: 0;
+	}
+
+	.tabLink {
+		white-space: nowrap;
+	}
+
+	> nav {
+		box-shadow: 0 0 0 black;
+	}
+}
